Central releationship id and reorganise on save update to fix drawing part collision and id ordering TODO: Add more reorg group like charts, tables on future updates

+++ cs/openxml_office/spreadsheet/Parts/2007/Worksheet.cs
@@ -6,9 +6,9 @@ namespace draviavemal.openxml_office.spreadsheet_2007
 
     public class Worksheet
     {
-        private readonly IntPtr ffiWorksheet;
+        private readonly ulong ffiWorksheet;
 
-        public Worksheet(IntPtr ffiWorksheet)
+        public Worksheet(ulong ffiWorksheet)
         {
             this.ffiWorksheet = ffiWorksheet;
         }
